Completed Value Added Features and Refactor

+++ HotelListing.API/Controllers/CountriesV2Controller.cs
@@ -1,45 +1,41 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
+using HotelListing.API.Data;
 using HotelListing.API.Core.Models.Country;
 using AutoMapper;
 using HotelListing.API.Core.Contracts;
 using Microsoft.AspNetCore.Authorization;
-using HotelListing.API.Core.Models;
+using HotelListing.API.Core.Exceptions;
+using Microsoft.AspNetCore.OData.Query;
 
 namespace HotelListing.API.Controllers
 {
     [Route("api/v{version:apiVersion}/countries")]
     [ApiController]
-    [ApiVersion("1.0", Deprecated = true)]
+    [ApiVersion("2.0")]
 
-    public class CountriesController : ControllerBase
+    public class CountriesV2Controller : ControllerBase
     {
         private readonly ICountriesRepository _countriesRepository;
-        private readonly ILogger<CountriesController> _logger;
+        private readonly ILogger<CountriesV2Controller> _logger;
         private readonly IMapper _mapper;
 
-        public CountriesController(IMapper mapper, ICountriesRepository countriesRepository, ILogger<CountriesController> logger)
+        public CountriesV2Controller(IMapper mapper, ICountriesRepository countriesRepository, ILogger<CountriesV2Controller> logger)
         {
             this._countriesRepository = countriesRepository;
             this._logger = logger;
             this._mapper = mapper;
         }
 
-        // GET: api/Countries/GetAll
-        [HttpGet("GetAll")]
-        public async Task<ActionResult<PageResult<GetCountryDto>>> GetCountries()
+        // GET: api/Countries
+        [HttpGet]
+        [EnableQuery]
+        public async Task<ActionResult<IEnumerable<GetCountryDto>>> GetCountries()
         {
             // Select * from Countries
-            var countries = await _countriesRepository.GetAllAsync<GetCountryDto>();
-            return Ok(countries);
-        }
-
-        // GET: api/Countries/?StartIndex=0&pagesize=25&PageNumber=1
-        [HttpGet]
-        public async Task<ActionResult<PageResult<GetCountryDto>>> GetPagedCountries([FromQuery] QueryParameters queryParameters)
-        {
-            var pagedCountriesResult = await _countriesRepository.GetAllAsync<GetCountryDto>(queryParameters);
-            return Ok(pagedCountriesResult);
+            var countries = await _countriesRepository.GetAllAsync();
+            var records = _mapper.Map<List<GetCountryDto>>(countries);
+            return Ok(records);
         }
 
         // GET: api/Countries/5
@@ -47,7 +43,15 @@ namespace HotelListing.API.Controllers
         public async Task<ActionResult<CountryDto>> GetCountry(int id)
         {
             var country = await _countriesRepository.GetDetails(id);
-            return Ok(country);
+
+            if (country == null)
+            {
+                throw new NotFoundException(nameof(GetCountries), id);
+            }
+
+            var countryDto = _mapper.Map<CountryDto>(country);
+
+            return Ok(countryDto);
         }
 
         // PUT: api/Countries/5
@@ -61,9 +65,18 @@ namespace HotelListing.API.Controllers
                 return BadRequest("Invalid Record Id");
             }
 
+            var country = await _countriesRepository.GetAsync(id);
+
+            if (country == null)
+            {
+                throw new NotFoundException(nameof(GetCountries), id);
+            }
+
+            _mapper.Map(updateCountryDto, country);
+
             try
             {
-                await _countriesRepository.UpdateAsync(id, updateCountryDto);
+                await _countriesRepository.UpdateAsync(country);
             }
             catch (DbUpdateConcurrencyException)
             {
@@ -84,10 +97,19 @@ namespace HotelListing.API.Controllers
         // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
         [HttpPost]
         [Authorize]
-        public async Task<ActionResult<CountryDto>> PostCountry(CreateCountryDto createCountry)
+        public async Task<ActionResult<Country>> PostCountry(CreateCountryDto createCountry)
         {
-            var country = await _countriesRepository.AddAsync<CreateCountryDto, CountryDto>(createCountry);
-            return CreatedAtAction(nameof(GetCountry), new { id = country.Id }, country);
+            //var countryOld = new Country
+            //{
+            //    Name = createCountry.Name,
+            //    ShortName = createCountry.ShortName
+            //};
+
+            var country = _mapper.Map<Country>(createCountry);
+
+            await _countriesRepository.AddAsync(country);
+
+            return CreatedAtAction("GetCountry", new { id = country.Id }, country);
         }
 
         // DELETE: api/Countries/5
@@ -95,6 +117,13 @@ namespace HotelListing.API.Controllers
         [Authorize(Roles = "Administrator")]
         public async Task<IActionResult> DeleteCountry(int id)
         {
+            var country = await _countriesRepository.GetAsync(id);
+
+            if (country == null)
+            {
+                throw new NotFoundException(nameof(GetCountries), id);
+            }
+
             await _countriesRepository.DeleteAsync(id);
             return NoContent();
         }
